Refactoring of database code
- Moved persistence.rs to bughouse_console to reduce scope of where it's
  built.
- Split database interactions to interface (persistence.rs) and
  implementation (database.rs). They now implement both reading and
  writing.
- Renamed SqlxServerHooks to DatabaseServerHooks as they are now
  agnostic of persistence method.
